Move X-Amzn-Query-Mode header setup to always be included if the model has the awsQueryCompatible trait
